Fix SRF05 formula
Use floats instead of fixed point math

+++ Application/source/distance.c
@@ -17,7 +17,7 @@
 void Distance_Task(void *arg)
 {
     (void)arg;
-    uint32_t  distanceCm;
+    float     distanceCm;
     bool      status;
 
     SRF05_Init();
